accuracy bugfix

+++ prediction_models/bow_lstm.py
@@ -166,9 +166,9 @@ class Model:
                 tf.argmax(self.targets, axis=2, output_type=tf.int32),
                 tf.argmax(self.prediction, axis=2, output_type=tf.int32))
             #correct = tf.cast(correct, tf.float32)
-            self.mask = tf.sequence_mask(self.seq_lengths, dtype=tf.int32)
+            self.mask = tf.sequence_mask(self.seq_lengths)
             tf.cast(correct, dtype=tf.int32)
-            correct *= self.mask
+            correct *= tf.cast(self.mask, dtype=tf.int32)
             self.correct = correct
             # Average over actual sequence lengths.
             correct = tf.reduce_sum(correct, reduction_indices=1)
